Typo Update README.md
There is a typo in your text in the "Examples" section. Specifically:

"Inside the examples folder, there two different examples,"

The correct version:

"Inside the examples folder, there are two different examples,"

The word "are" is used instead of "is" to agree with the plural noun ("two different examples").

Corrected.